Add files via upload
Up untill now two maps, and simple collisions were made, The movement is not definitive as it is only for debug porpuses and later on it should be created by the LLM at runtime. Everything seems to work fluidly if the window is not scaled, when trying to change the size the collision move unexpectedly for now.